Changed layout of MainActivity to have a now playing bar at button of layout to display currently playing song's album art and song and album title, as well as buttons for pausing and playing.
Added new IntentService to be used to play media files in the background now instead of an AsyncTask. Uses a static reference to the media player object to allow access to the media player throughout the entire application. Also changed how the song information is displayed in NowPlayingActivity to work together with the new IntentService.