Changed the Main Activity to correctly display the currently playing media information. Added a touch listener to the bottom bar to go the Now Playing activity when touched.
Updated MediaPlayerService to properly check if there is a media player object instantiated yet. If there is, there is a check to see if the current song is the same as the selected song. This will allow the media player to know if it needs to be reset or not.

Now Playing activity will now change based on the current song and the media player will only reset when a new song is selected.